refactor(storekit): a predicate leaf's operand validation has its own file
predicate.go carries the filter grammar and the compiler walk;
predicateoperand.go carries how one leaf's value is checked against its
field type and turned into a bind parameter (comparisonSQL, existsOperand,
inOperand, scalarOperand, escapeLike) — a pure move, no behaviour change.
